requirement.txt
# Backend dependencies
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
mysqlclient==2.2.6
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
